+++ src/userspace/libc/stdlib/malloc.c
@@ -1,7 +1,7 @@
 #include <stdint.h>
 #include <syscall.h>
 
-void *malloc(u8 size)
+void *malloc(u32 size)
 {
 	return (void *)syscall_malloc(size);
 }
