+++ src/userspace/libc/string/strcmp.c
@@ -6,7 +6,7 @@ char strcmp(char *a, char *b)
 	if (strlen(a) != strlen(b))
 		return 1;
 
-	for (u8 i = 0; i < strlen(a); i++)
+	for (u32 i = 0; i < strlen(a); i++)
 		if (a[i] != b[i])
 			return 1;
