+++ src/userspace/libc/string/strcpy.c
@@ -3,9 +3,9 @@
 
 void strcpy(char *dest, char *orig)
 {
-	u8 s_orig = strlen(orig);
+	u32 s_orig = strlen(orig);
 
-	for (u8 i = 0; i < s_orig; i++)
+	for (u32 i = 0; i < s_orig; i++)
 		dest[i] = orig[i];
 	dest[s_orig] = 0;
 }
